Update Category entity Update CategoryForm model Update CategoryManageService class

+++ shop/services/manage/shop/CategoryManageService.php
@@ -4,6 +4,8 @@ namespace shop\services\manage\shop;
 
 use shop\entities\Meta;
 use shop\entities\shop\Category;
+use shop\entities\shop\product\Product;
+use shop\exceptions\NotFoundException;
 use shop\forms\manage\shop\CategoryForm;
 use yii\helpers\Inflector;
 
@@ -11,7 +13,7 @@ class CategoryManageService
 {
     public function create(CategoryForm $form) : Category
     {
-        $parent = Category::findOne(['id' => $form->parentId]);
+        $parent = $this->getCategory($form->parentId);
 
         $category = Category::create(
             $form->name,
@@ -33,7 +35,7 @@ class CategoryManageService
 
     public function edit($id, CategoryForm $form)
     {
-        $category = Category::findOne(['id' => $id]);
+        $category = $this->getCategory($id);
 
         $this->assertIsNotRoot($category);
 
@@ -50,7 +52,7 @@ class CategoryManageService
         );
 
         if ($form->parentId !== $category->parent->id) {
-            $parent = Category::findOne(['id' => $form->parentId]);
+            $parent = $this->getCategory($form->parentId);
 
             $category->appendTo($parent);
         }
@@ -58,11 +60,41 @@ class CategoryManageService
         $category->save();
     }
 
-    public function remove($id)
+    public function moveUp($id)
     {
-        $category = Category::findOne(['id' => $id]);
+        $category = $this->getCategory($id);
 
         $this->assertIsNotRoot($category);
+
+        if ($prev = $category->prev) {
+            $category->insertBefore($prev);
+        }
+
+        $category->save();
+    }
+
+    public function moveDown($id)
+    {
+        $category = $this->getCategory($id);
+
+        $this->assertIsNotRoot($category);
+
+        if ($next = $category->next) {
+            $category->insertAfter($next);
+        }
+
+        $category->save();
+    }
+
+    public function remove($id)
+    {
+        $category = $this->getCategory($id);
+
+        $this->assertIsNotRoot($category);
+
+        if (Product::find()->andWhere(['category_id' => $id])->exists()) {
+            throw new \DomainException('Unable to delete category with products.');
+        }
 
         $category->delete();
     }
@@ -72,5 +104,14 @@ class CategoryManageService
         if ($category->isRoot()) {
             throw new \DomainException('Unable to manage the root category.');
         }
+    }
+
+    private function getCategory($id) : Category
+    {
+        if (!$category = Category::findOne(['id' => $id])) {
+            throw new NotFoundException('Category not found.');
+        }
+
+        return $category;
     }
 }
